Rename contracts package to contractgateway
